Added primaryKey introspection and made sync and dbPull functions dry Co-authored-by:	Ben Engelberg <benjaminsengelberg@gmail.com>
