+++ src/functions/sync.ts
@@ -1,54 +1,12 @@
-import { ConnectDb, DisconnectDb } from './Db.ts'
-import { tableListQuery, tableConstQuery, columnInfoQuery } from '../queries/introspection.ts'
+import { introspect } from './introspect.ts';
 
 // take the data from the model.ts file and reverse engineer it
 // essentially make it look like the query results
-
-// INTERFACES
-interface ITableQueryRecords {
-    table_name: string
-}
-
-interface IColumnQueryRecords {
-    schemaname: string
-    table_name: string
-    column_name: string
-    column_type: string
-    col_default: unknown
-    not_null: boolean
-}
 
 interface ModelInfo {
     table_name: string
     columns: Record<string, Record<string, unknown>>
 }
-
-// TYPE GUARD FUNCTIONS
-const recordObjectType = (record: object): record is ITableQueryRecords => {
-    return 'table_name' in record;
-}
-
-const colRecordObjectType = (record: object): record is IColumnQueryRecords => {
-    return 'schemaname' in record && 'table_name' in record && 'column_name' in record &&
-        'column_type' in record && 'col_default' in record && 'not_null' in record
-}
-
-const introspectDb = async () => {
-    const db = await ConnectDb();
-
-    const tableList = await db.queryObject(tableListQuery);
-    const columnList = await db.queryObject(columnInfoQuery);
-    const constraintList = await db.queryObject(tableConstQuery);
-
-    DisconnectDb(db);
-
-    return {
-        tableList: tableList.rows,
-        columnList: columnList.rows,
-        constraintList: constraintList.rows
-    }
-}
-
 
 const jsonify = (text: string): ModelInfo[] => {
         // remove extraneous text
@@ -89,8 +47,7 @@ const createTable = (table_name: string, columns: Record<string, Record<string, 
 }
 
 export const sync = async () => {
-    const { tableList, columnList, constraintList } = await introspectDb();
-
+    const tableListObj = await introspect();
     // import typescript model of tables
     let modelText = Deno.readTextFileSync('./src/functions/temp_model.ts');
 
@@ -99,30 +56,7 @@ export const sync = async () => {
     let createTableQueries = ``;
     let alterTableQueries = ``;
 
-    // convert table list to an object
-    const tableListObj: Record<string, Record<string, Record<string, unknown>>> = {};
-
-    // Add each table to the tableListObj for easier access
-    tableList.forEach(el => {
-        if(typeof el === 'object' && el !== null && 'table_name' in el){
-            if(recordObjectType(el)) tableListObj[String(el.table_name)] = {};
-        }
-    });
-
-    // Add columns to the table's object in the tableListObj for easier look-up
-    columnList.forEach(el => {
-        if(typeof el === 'object' && el !== null && colRecordObjectType(el)){
-            tableListObj[el.table_name][el.column_name] = {};
-
-            const refObj = tableListObj[el.table_name][el.column_name];
-
-            refObj['column_type'] = el.column_type;
-            refObj['col_default'] = el.col_default;
-            refObj['not_null'] = el.not_null;
-    }
-    })
-
-    console.log(tableListObj, modelArray)
+    console.log(tableListObj)
 
     modelArray.forEach(el => {
         // SQL statements for tables not currently in the database
@@ -143,5 +77,7 @@ export const sync = async () => {
         }
     })
 
-    console.log(createTableQueries)
+    //console.log(createTableQueries)
 }
+
+sync()
